+++ addons/scene_manager/scenes.gd
@@ -5,4 +5,4 @@
 #
 extends Node
 
-var scenes: Dictionary = {"_ignore_list":["res://addons"],"_sections":["Character","Menu"],"scene1":{"sections":["Menu"],"value":"res://demo/scene1.tscn"},"scene2":{"sections":[],"value":"res://demo/scene2.tscn"},"scene3":{"sections":["Menu","Character"],"value":"res://demo/scene3.tscn"}}
+var scenes: Dictionary = {"_ignore_list":["res://addons"],"_sections":["Character","Menu"],"loading":{"sections":[],"value":"res://demo/loading.tscn"},"scene1":{"sections":["Menu"],"value":"res://demo/scene1.tscn"},"scene2":{"sections":[],"value":"res://demo/scene2.tscn"},"scene3":{"sections":["Menu","Character"],"value":"res://demo/scene3.tscn"}}